Add support for Firefox (76+).
Firefox recently added support for AudioWorklets. kilobeat still wasn't
working in it due to the cycle in the graph where the ChannelMerger fed
back into CustomNodes. Per the spec, cycles are muted; not sure why that
wasn't happening in Chrome.

Now there's a delay node following the ChannelMerger, eliminating the
cycle, and this runs in Firefox. (Tested playing offline and loading a
recording.)